requirements.txt
fastapi
python-multipart
pdf2image
boto3
uvicorn
pillow
python-dotenv